Function lishp::lexer::tokenize

pub fn tokenize<T: Into<String>>(src: T) -> Result<Vec<Token>, InvalidTokenError>

Turn some source code into a list of Tokens.

This is pretty much just a thin wrapper around the Lexer struct. It creates a lexer then keeps reading tokens until there aren't any left, skipping any whitespace.