Struct lishp::parser::Parser

pub struct Parser { /* fields omitted */ }

The Parser.

Examples

Methods

impl Parser

fn new(tokens: Vec<Token>) -> Parser

Create a new Parser using a list of Tokens.

fn peek(&self) -> Option<&Token>

Look at the next Token, but don't consume it.

fn next(&mut self) -> Option<&Token>

Consume the next token in the stream.

fn parse(&mut self) -> LishpResult<Type>

Do the actual parsing and get the resultant AST.