Trait lishp::visitor::Visitor

pub trait Visitor {
    fn visit(&mut self, node: &mut Type) { ... }
    fn visit_list(&mut self, node: &mut Type) { ... }
    fn visit_atom(&mut self, node: &mut Type) { ... }
    fn visit_boolean(&mut self, b: &mut Type) { ... }
    fn visit_integer(&mut self, i: &mut Type) { ... }
    fn visit_float(&mut self, f: &mut Type) { ... }
    fn visit_string(&mut self, s: &mut Type) { ... }
    fn visit_symbol(&mut self, s: &mut Type) { ... }
}

The generic Visitor trait. Anything implementing this trait will be able to visit the nodes in an AST.

Note that all visit_*() methods take in a mutable reference to a Type. This means that a visitor can change literally anything in the AST if they want to, making it easy for you to add in arbitrary optimiser passes or "compiler plugins" before the actual code is run.

Examples

use lishp::visitor::Visitor;
struct MyVisitor;

// simply writing `impl Visitor for MyVisitor {}` would give us access to
// all the `Visitor` visit_*() functions. Therefore, you only need to
// override the methods you need.
impl Visitor for MyVisitor {
  fn visit_symbol(&mut self, s: &mut Type) {
    println!("Visiting symbol: {:?}", s);
  }
}Run

Provided Methods

fn visit(&mut self, node: &mut Type)

The default behaviour is to delegate to either visit_list() or visit_atom() depending on what type of AST node it is.

fn visit_list(&mut self, node: &mut Type)

Just recursively visit each node in the list.

fn visit_atom(&mut self, node: &mut Type)

Visiting an atom simply delegates to the appropriate visitor for that node type (visit_boolean(), visit_integer(), etc).

fn visit_boolean(&mut self, b: &mut Type)

Visit a boolean. Default behaviour is a no-op.

fn visit_integer(&mut self, i: &mut Type)

Visit an integer. Default behaviour is a no-op.

fn visit_float(&mut self, f: &mut Type)

Visit a float. Default behaviour is a no-op.

fn visit_string(&mut self, s: &mut Type)

Visit a string. Default behaviour is a no-op.

fn visit_symbol(&mut self, s: &mut Type)

Visit a symbol. Default behaviour is a no-op.